Refactor asyncResourceBundle into different files
