+++ rule-audit-client/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  moduleFileExtensions: ['ts', 'tsx', 'js'],
+  transform: {
+    '^.+\\.(ts|tsx|js)$': 'ts-jest'
+  }
+};
