+++ rule-audit-client/src/redux/utils/createAsyncResourceBundle/index.ts
@@ -1,164 +1,27 @@
-import without from 'lodash/without';
-import isEqual from 'lodash/isEqual';
+import isEqual from "lodash/isEqual";
 
-interface BaseResource {
+import createSelectors from "./createSelectors";
+import createActions, {
+  Actions,
+  FETCH_START,
+  FETCH_SUCCESS,
+  FETCH_SUCCESS_IGNORE,
+  FETCH_ERROR,
+  UPDATE_START,
+  UPDATE_SUCCESS,
+  UPDATE_ERROR
+} from "./createActions";
+import {
+  formatIncomingResourceData,
+  applyStatusIds,
+  removeStatusIds,
+  getOrderFromIncomingResourceData,
+  getStatusIdsFromData,
+  globalLoadingIndicator
+} from "./utils";
+
+export interface BaseResource {
   id: string;
-}
-
-const FETCH_START = 'FETCH_START';  
-const FETCH_SUCCESS = 'FETCH_SUCCESS';
-const FETCH_SUCCESS_IGNORE = 'FETCH_SUCCESS_IGNORE'; // clears loading ids for unchanged collections during collection polling
-const FETCH_ERROR = 'FETCH_ERROR';
-const UPDATE_START = 'UPDATE_START';
-const UPDATE_SUCCESS = 'UPDATE_SUCCESS';
-const UPDATE_ERROR = 'UPDATE_ERROR';
-
-interface FetchStartAction {
-  entity: string;
-  type: 'FETCH_START';
-  payload: { ids?: string[] | string };
-}
-
-interface FetchSuccessAction<Resource> {
-  entity: string;
-  type: 'FETCH_SUCCESS';
-  payload: {
-    data: Resource | Resource[] | any;
-    order?: string[];
-    pagination?: IPagination;
-    time: number;
-  };
-}
-
-interface FetchSuccessIgnoreAction<Resource> {
-  entity: string;
-  type: 'FETCH_SUCCESS_IGNORE';
-  payload: {
-    data: Resource | Resource[] | any;
-    time: number;
-  };
-}
-
-interface FetchErrorAction {
-  entity: string;
-  type: 'FETCH_ERROR';
-  payload: {
-    error: string;
-    time: number;
-    ids?: string | string[];
-  };
-}
-
-interface UpdateStartAction<Resource> {
-  entity: string;
-  type: 'UPDATE_START';
-  payload: { id?: string | string; data: Resource | any };
-}
-
-interface UpdateSuccessAction<Resource> {
-  entity: string;
-  type: 'UPDATE_SUCCESS';
-  payload: { data: Resource | any; id: string; time: number };
-}
-
-interface UpdateErrorAction {
-  entity: string;
-  type: 'UPDATE_ERROR';
-  payload: {
-    error: string;
-    id: string;
-    time: number;
-  };
-}
-
-type Actions<Resource> =
-  | FetchStartAction
-  | FetchSuccessAction<Resource>
-  | FetchSuccessIgnoreAction<Resource>
-  | FetchErrorAction
-  | UpdateStartAction<Resource>
-  | UpdateSuccessAction<Resource>
-  | UpdateErrorAction;
-
-const defaultArray = [] as string[];
-const globalLoadingIndicator = '@@ALL';
-
-const getStatusIdsFromData = (
-  data: BaseResource | BaseResource[] | any
-): string[] | string =>
-  data instanceof Array
-    ? data.map((resource: BaseResource) => resource.id || '')
-    : data.id || '';
-
-const applyStatusIds = (
-  currentIds: string[],
-  incomingIds?: string | string[]
-) => currentIds.concat(incomingIds || globalLoadingIndicator);
-
-const removeStatusIds = (
-  currentIds: string[],
-  incomingIds: string[] | string = ''
-): string[] =>
-  incomingIds instanceof Array
-    ? without(currentIds, ...incomingIds, globalLoadingIndicator)
-    : without<string>(currentIds, incomingIds);
-
-function formatIncomingResourceData<Resource extends BaseResource>(
-  data: { [id: string]: Resource } | {},
-  newData: Resource | Resource[],
-  resourceName: string
-): Resource | { [id: string]: Resource } {
-  if (newData instanceof Array) {
-    const result: { [id: string]: Resource } = {
-      ...data,
-      ...newData.reduce((acc, model: BaseResource, index) => {
-        if (!model.id) {
-          throw new Error(
-            `[asyncResourceBundle]: Cannot apply new data - incoming resource ${resourceName} is missing ID at index ${index}.`
-          );
-        }
-        return {
-          ...acc,
-          [model.id]: model
-        };
-      }, {})
-    };
-    return result;
-  }
-
-  if (!newData.id) {
-    throw new Error(
-      `[asyncResourceBundle]: Cannot apply new data - incoming resource ${resourceName} with keys ${Object.keys(
-        newData
-      ).join(', ')} is missing id.`
-    );
-  }
-
-  return {
-    ...data,
-    [newData.id]: newData
-  };
-}
-
-function getOrderFromIncomingResourceData<Resource extends BaseResource>(
-  newData: Resource | Resource[],
-  resourceName: string,
-  currentOrder: string[] = defaultArray,
-  newOrder?: string[]
-): string[] {
-  const order =
-    newOrder ||
-    (newData instanceof Array
-      ? (newData as Resource[]).map((model, index) => {
-          if (!model.id) {
-            throw new Error(
-              `[asyncResourceBundle]: Cannot apply new data - incoming resource ${resourceName} is missing ID at index ${index}.`
-            );
-          }
-          return model.id;
-        })
-      : []);
-  return isEqual(currentOrder, order) ? currentOrder : order;
 }
 
 interface IPagination {
@@ -166,7 +29,7 @@ interface IPagination {
   totalPages: number;
   currentPage: number;
 }
-interface State<Resource> {
+export interface State<Resource> {
   data: Resource | { [id: string]: Resource } | any;
   pagination: IPagination | null;
   lastError: string | null;
@@ -214,37 +77,7 @@ function createAsyncResourceBundle<Resource, Action>(
   const { indexById } = options;
   const selectLocalState = options.selectLocalState
     ? options.selectLocalState
-    : (state: any): State<Resource> => state[entityName];
-
-  const selectPagination = (state: RootState) =>
-    selectLocalState(state).pagination;
-
-  const selectCurrentError = (state: RootState) =>
-    selectLocalState(state).error;
-
-  const selectLastError = (state: RootState) =>
-    selectLocalState(state).lastError;
-
-  const selectLastFetch = (state: RootState) =>
-    selectLocalState(state).lastFetch;
-
-  const selectIsLoading = (state: RootState) =>
-    !!selectLocalState(state).loadingIds.length;
-
-  const selectIsLoadingById = (state: RootState, id: string) =>
-    selectLocalState(state).loadingIds.indexOf(id) !== -1;
-
-  const selectById = (state: RootState, id: string): Resource | undefined =>
-    selectLocalState(state).data[id];
-
-  const selectIsLoadingInitialDataById = (state: RootState, id: string) =>
-    !selectById(state, id) &&
-    selectLocalState(state).loadingIds.indexOf(id) !== -1;
-
-  const selectLastFetchOrder = (state: RootState): string[] =>
-    selectLocalState(state).lastFetchOrder || defaultArray;
-
-  const selectAll = (state: RootState) => selectLocalState(state).data;
+    : (state: RootState): State<Resource> => state[entityName];
 
   const initialState: State<Resource> = {
     data: options.initialData || {},
@@ -255,59 +88,6 @@ function createAsyncResourceBundle<Resource, Action>(
     loadingIds: [],
     updatingIds: []
   };
-
-  const fetchStartAction = (ids?: string[] | string): FetchStartAction => ({
-    entity: entityName,
-    type: FETCH_START,
-    payload: { ids }
-  });
-
-  const fetchSuccessAction = (
-    data: Resource | Resource[] | any,
-    { pagination, order }: { pagination?: IPagination; order?: string[] } = {}
-  ): FetchSuccessAction<Resource> => ({
-    entity: entityName,
-    type: FETCH_SUCCESS,
-    payload: { data, pagination, order, time: Date.now() }
-  });
-
-  const fetchSuccessIgnoreAction = (
-    data: Resource | Resource[] | any
-  ): FetchSuccessIgnoreAction<Resource> => ({
-    entity: entityName,
-    type: FETCH_SUCCESS_IGNORE,
-    payload: { data, time: Date.now() }
-  });
-
-  const fetchErrorAction = (
-    error: string,
-    ids?: string | string[]
-  ): FetchErrorAction => ({
-    entity: entityName,
-    type: FETCH_ERROR,
-    payload: { error, ids, time: Date.now() }
-  });
-
-  const updateStartAction = (data: Resource): UpdateStartAction<Resource> => ({
-    entity: entityName,
-    type: UPDATE_START,
-    payload: { data }
-  });
-
-  const updateSuccessAction = (
-    id: string,
-    data?: Resource
-  ): UpdateSuccessAction<Resource> => ({
-    entity: entityName,
-    type: UPDATE_SUCCESS,
-    payload: { id, data, time: Date.now() }
-  });
-
-  const updateErrorAction = (error: string, id: string): UpdateErrorAction => ({
-    entity: entityName,
-    type: UPDATE_ERROR,
-    payload: { error, id, time: Date.now() }
-  });
 
   const isAction = (
     action: Actions<Resource> | Action
@@ -460,29 +240,10 @@ function createAsyncResourceBundle<Resource, Action>(
       updateSuccess: UPDATE_SUCCESS,
       updateError: UPDATE_ERROR
     },
-    actions: {
-      fetchStart: fetchStartAction,
-      fetchSuccess: fetchSuccessAction,
-      fetchSuccessIgnore: fetchSuccessIgnoreAction,
-      fetchError: fetchErrorAction,
-      updateStart: updateStartAction,
-      updateSuccess: updateSuccessAction,
-      updateError: updateErrorAction
-    },
-    selectors: {
-      selectPagination,
-      selectCurrentError,
-      selectLastError,
-      selectLastFetch,
-      selectIsLoading,
-      selectIsLoadingById,
-      selectIsLoadingInitialDataById,
-      selectById,
-      selectLastFetchOrder,
-      selectAll
-    }
+    actions: createActions(entityName),
+    selectors: createSelectors<Resource, RootState>(selectLocalState)
   };
 }
 
-export { Actions, State, IPagination, globalLoadingIndicator };
+export { IPagination, globalLoadingIndicator };
 export default createAsyncResourceBundle;
